[Decomp][Item] Cohesion split: extract persistence to ItemPersistence.cpp
